pyproject.toml
[project]
name = "nj-crashes"
version = "0.1.0"
description = "Analysis of NJ State Police fatal crash data and NJDOT crash data"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Ryan Williams" }
]
dependencies = [
    "atproto>=0.0.62",
    "beautifulsoup4>=4.13",
    "click>=8.1",
    "dask>=2024",
    "dotenv>=0.9.9",
    "dvc[s3]>=3.59",
    "geopandas>=1.0",
    "GitPython>=3.1",
    "humanize>=4.12",
    "joblib>=1.5",
    "juq-py>=0.5.2",
    "jupyterlab>=4.4",
    "lxml>=5.3",
    "numpy>=2.2",
    "pandas>=2.2",
    "papermill>=2.6",
    "plotly>=6.0",
    "pyarrow>=20.0",
    "PyGithub>=2.6",
    "pyogrio>=0.10",
    "pyproj>=3.7",
    "requests>=2.32",
    "s3fs>=2024",
    "shapely>=2.1",
    "simplejson>=3.20",
    "slack-sdk>=3.37.0",
    "SQLAlchemy>=2.0",
    "tabula-py>=2.10",
    "utz>=0.20",
]

[project.optional-dependencies]
bsky = [
    "atproto>=0.0.55",
    "python-dotenv>=1.0",
]
slack = [
    "slack-sdk>=3.33",
    "python-dotenv>=1.0",
    "stdlb>=0.0.4",
]

[project.scripts]
njsp = "njsp.cli.main:main"
njdot = "njdot.cli.main:main"
rawdata = "njdot.rawdata.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nj_crashes", "njsp", "njdot"]
